Helmet implement

+++ src/Pages/MyJobs/Myjobs.jsx
@@ -5,6 +5,7 @@ import Navbar from "../../Components/Shared/Navbar";
 import img1 from "../../assets/Images/animatepng.png";
 import Footer from "../../Components/Shared/Footer/Footer";
 import Swal from "sweetalert2";
+import { Helmet } from "react-helmet-async";
 
 const Myjobs = () => {
   const { user } = useContext(AuthContext);
@@ -19,38 +20,47 @@ const Myjobs = () => {
   const handelDelete=(_id)=>{
     console.log(_id);
 
-    fetch(`http://localhost:5000/jobs/${_id}`, {
-        method: "DELETE",
-      })
-
-.then((res) => res.json())
- 
-.then((data) => {
-    if (data.deletedCount > 0) {
-      Swal.fire({
-        title: "Are you sure?",
-        text: "You won't be able to revert this!",
-        icon: "warning",
-        showCancelButton: true,
-        confirmButtonColor: "#3085d6",
-        cancelButtonColor: "#d33",
-        confirmButtonText: "Yes, delete it!"
-      })
     Swal.fire({
-      title: "Deleted!",
-      text: "Your file has been deleted.",
-      icon: "success"
-     
-    });
-    const remaing = jobsstate.filter((jobdelete)=> jobdelete._id !==_id)
-    setJobsState(remaing)
-  }
- 
-});
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result)=>{
+      if(result.isConfirmed){
+        fetch(`http://localhost:5000/jobs/${_id}`, {
+          method: "DELETE",
+        })
+  
+  .then((res) => res.json())
+   
+  .then((data) => {
+      if (data.deletedCount > 0) {
+        
+      Swal.fire({
+        title: "Deleted!",
+        text: "Your file has been deleted.",
+        icon: "success"
+       
+      });
+      const remaing = jobsstate.filter((jobdelete)=> jobdelete._id !==_id)
+      setJobsState(remaing)
+    }
+   
+  });
+      }
+    })
+
+   
 
   }
   return (
     <div>
+      <Helmet>
+        <title>JOBS MARKET | MY JOBS</title>
+      </Helmet>
       <Navbar></Navbar>
       <div
         style={{
